refactor: improve streaming architecture with helper extraction and type safety
- Extract prepareMessages and handleStreamFinish helper functions from create-chat-stream-response.ts
- Reduce main function complexity from ~290 lines to ~100 lines for better maintainability
- Add StreamContext interface for better state management
- Improve type safety in DataSection component by replacing 'any' with proper union types
- Add DataPart and DataRelatedQuestionsPart type definitions
- Better separation of concerns and code organization
- Enhanced readability and testability of streaming components
