+++ components/related-questions.tsx
@@ -55,7 +55,7 @@ export const RelatedQuestions: React.FC<RelatedQuestionsProps> = ({
                   <ArrowRight className="h-4 w-4 mr-2 mt-1.5 flex-shrink-0 text-accent-foreground/50" />
                   <Button
                     variant="link"
-                    className="flex-1 justify-start px-0 py-1 h-fit font-semibold text-accent-foreground/50 whitespace-normal text-left"
+                    className="flex-1 justify-start px-0 py-0 h-fit font-semibold text-accent-foreground/50 whitespace-normal text-left"
                     type="submit"
                     name={'related_query'}
                     value={item.question}
