+++ lib/streaming/helpers/types.ts
@@ -1,0 +1,9 @@
+export interface StreamContext {
+  chatId: string
+  userId: string
+  modelId: string
+  messageId?: string
+  trigger?: string
+  initialChat: any
+  abortSignal?: AbortSignal
+}
